+++ src/main/java/ru/clevertec/core/parser/temp/ParserValue.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core;
+package ru.clevertec.core.parser.temp;
 
 import ru.clevertec.core.node.Node;
