+++ src/main/java/ru/clevertec/core/parser/temp/ReplacerComma.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core;
+package ru.clevertec.core.parser.temp;
 
 import java.util.UUID;
