+++ src/test/java/ru/clevertec/ConverterTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import ru.clevertec.core.AbstractContainer;
-import ru.clevertec.core.Parser2;
+import ru.clevertec.core.parser.Parser2;
 import ru.clevertec.core.node.NodeFactory;
 import ru.clevertec.core.service.deserialization.NodeConverterFactory;
 import ru.clevertec.core.service.serialization.ConverterFactoryImpl;
